+++ pakyow-core/CHANGELOG.md
@@ -4,6 +4,7 @@
     `Helpers` is included only into `CallContext`. Extensions to `App` is now
     considered an edge-case and should be done through other normal means.
   * IMPROVED hooks by removing `Helpers::Hooks` in favor of `Support::Hookable`.
+  * ADDED verbose log level.
 
 # 0.11.3